Add a CUDA backend alongside Metal
The GPU kernels were written against mx.fast.metal_kernel. They now go
through pydiffvg/gpu_backend.py, which compiles the same kernel body for
either Apple Metal (pydiffvg/metal/*.metal) or NVIDIA CUDA
(pydiffvg/cuda/*.cu, mx.fast.cuda_kernel) and picks the backend from
mx.metal.is_available() / mx.cuda.is_available().

Each kernel body is written once, in a backend-neutral dialect: plain C
plus DVG_THREAD_INDEX (thread index, and on CUDA the bounds check nvrtc
needs), DVG_ADD / DVG_STORE (atomics), DVG_XYZ (float4 -> float3) and
DVG_CEIL. Nothing else in a body differs between backends. The device
helpers are translated per backend, so pydiffvg/cuda/ mirrors
pydiffvg/metal/ file for file; pydiffvg/cuda/README-porting.md records
what the translation has to preserve.

Bit parity with the C++ core needs different tools on each backend:
Metal takes `#pragma METAL fp contract(off)' and a correctly rounded
sqrt correction, while nvrtc offers no contraction pragma and needs the
__fmul_rn / __fadd_rn intrinsics, its sqrtf already being correctly
rounded. scene_gpu grew a scene.metal / scene.cu pair for its five
helper kernels and a supports_backend() so an unsupported backend falls
back to the C++ scene build.

compute_distance.h: cast fabs() and sqrt() results to float. With
<math.h> in scope they return double, and max(float, double) does not
compile under g++, which clang accepts. This is what makes the extension
build on Linux.

Verified on Apple Silicon: sampled rendering is bit-identical to the C++
core, prefiltered rendering agrees to 6.2e-06 with no pixel past 1e-4,
and the linear-loss gradient matches theory (251.33 against 2*pi*r =
251.327). The CUDA path builds and installs on Linux; its kernel-level
parity is not yet measured.

+++ compute_distance.h
@@ -365,7 +365,9 @@ float cubic_distance_lower_bound(const Vector2f &p0, const Vector2f &p1,
     auto h = max(point_to_segment_distance(p1, p0, p3), point_to_segment_distance(p2, p0, p3));
     auto m = 0.f;
     for (const Vector2f &q : {p0 - pt, p1 - pt, p2 - pt, p3 - pt}) {
-        m = max(m, max(fabs(q.x), fabs(q.y)));
+        // explicit float: with <math.h> in scope fabs() returns double, and
+        // max(float, double) does not compile (clang accepts it, g++ does not)
+        m = max(m, max(float(fabs(q.x)), float(fabs(q.y))));
     }
     return point_to_segment_distance(pt, p0, p3) - h - 1e-5f * (1 + m);
 }
@@ -648,8 +650,10 @@ bool compute_distance(const SceneData &scene,
         auto a11 = c(0, 1) * c(0, 1) + c(1, 1) * c(1, 1);
         auto half_tr = (a00 + a11) / 2;
         auto det = a00 * a11 - a01 * a01;
-        auto lambda_max = half_tr + sqrt(max(half_tr * half_tr - det, 0.f));
-        local_max_radius = max_radius * sqrt(max(lambda_max, 0.f));
+        // explicit float: with <math.h> in scope sqrt() returns double, and
+        // max(double, float) does not compile (clang accepts it, g++ does not)
+        auto lambda_max = half_tr + float(sqrt(max(half_tr * half_tr - det, 0.f)));
+        local_max_radius = max_radius * float(sqrt(max(lambda_max, 0.f)));
     }
 
     while (stack_size > 0) {
